requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.2
pydantic==1.8.2
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
